FIS-87 Update code for Login and Register
