+++ src/main/java/Login/services/LoginService.java
@@ -3,35 +3,12 @@ package Login.services;
 import Login.exceptions.IncorrectPassword;
 import Login.exceptions.UserDoesNotExistException;
 import ParentCode.Exceptions.EmptyField;
+import ParentCode.Service;
 import Register.actionMode.User;
-import Register.services.FileSystemService;
-import Register.services.UserService;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.commons.io.FileUtils;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.List;
 import static Register.services.UserService.encodePassword;
 
 
-public class LoginService {
-
-    private static List<User> users;
-    private static final Path USERS_PATH = FileSystemService.getPathToFile("config", "users.json");
-
-    public static void loadUsersFromFile() throws IOException {
-
-        if (!Files.exists(USERS_PATH)) {
-            FileUtils.copyURLToFile(UserService.class.getClassLoader().getResource("Register/users.json"), USERS_PATH.toFile());
-        }
-
-        ObjectMapper objectMapperClient = new ObjectMapper();
-
-        users = objectMapperClient.readValue(USERS_PATH.toFile(), new TypeReference<List<User>>() {
-        });
-    }
+public class LoginService extends Service {
 
     public static void verifyLogin (String username, String password) throws UserDoesNotExistException, IncorrectPassword, EmptyField {
