+++ src/main/java/Register/controllers/ManagerRegController.java
@@ -55,6 +55,7 @@ public class ManagerRegController extends Controller {
             UserService.addUserManager(libraryName.getText(), address.getText(), email.getText(), phoneNumber.getText(), username.getText(), password.getText());
             URL url = new File("src/main/resources/Login/Login.fxml").toURI().toURL();
             super.handle(event, url);
+            UserService.loadUsersFromFile();
         } catch (AlreadyExistsException e) {
             empty.setText(null);
             registrationMessage.setText(e.getMessage());
